MicrFinder changes to allow search on pages without OCR results. Changes to support for GdPicture engine (minus the actual GdPicture support)

+++ ReusableComponents/COMComponents/UCLIDRasterAndOCRMgmt/Core/Code/SpatialString_Internal.cpp
@@ -4842,8 +4842,8 @@ void CSpatialString::loadFromGoogleJson(const string& strInputFile, long nPageNu
 	long height = heightIt->value.GetInt();
 	bool hasNormalizedVertices = false;
 	bool hasVertices = false;
-	double widthConvertedFromPoints = width * 300 / 72;
-	double heightConvertedFromPoints = height * 300 / 72;
+	double widthConvertedFromPoints = (double)width * 300 / 72;
+	double heightConvertedFromPoints = (double)height * 300 / 72;
 
 	vector<CPPLetter> letters;
 	vector<double> thetas;
